updated streamlit app
